Always put set the `status` label on the `requests_total` metric.
We only set the `status` label for failed requests, for consistency
also set it when counting successful requests.

While at it update our metric context managers to ensure it create
metrics with the labels set when the context manager is created.
Although there could be some use case for this behavior it is just not
intuitive enough, so copy the labels when the context manager is
created.

Signed-off-by: Alban Bedel <albeu@free.fr>